+++ scripts/CommentBoilerplate_OutputFiles/Boilerplate.py
@@ -1,3 +1,0 @@
-# Correct Python comment.
-#   Missing space after marker.
-# print("Hello from Python")