+++ scripts/CommentBoilerplate_TestFiles/Boilerplate.c
@@ -1,2 +1,0 @@
-// C comment.
-char *s = "Hello";