+++ scripts/CommentBoilerplate_TestFiles/Boilerplate.cpp
@@ -1,2 +1,0 @@
-// C++ comment.
-int main() { return 0; }