+++ scripts/CommentBoilerplate_TestFiles/Boilerplate.ps1
@@ -1,2 +1,0 @@
-# PowerShell script comment.
-echo "Hello from PowerShell"