+++ scripts/CommentBoilerplate_TestFiles/Boilerplate.sh
@@ -1,2 +1,0 @@
-# Shell script comment.
-echo "Hello from shell"