+++ scripts/toml_insert.ps1
@@ -1,3 +1,22 @@
+<# 
+Package: pixi_pyprojSetup
+
+Copyright (C) 2025  Ben Hurwitz <bchurwitz+pixi_pyprojSetup> under GNU GPL v3.0.
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#>
+
 <# toml_insert.ps1
 
 .SYNOPSIS
@@ -16,7 +35,7 @@
     The line AFTER this string appears will be where the insertions take place. Be sure to pass a unique string here, e.g. 'versions = ' and not just 'versions' (which could be elsewhere). 
 
 .PARAMETER Debug  
-    A switch that, if specified, turns on debugging messages. 
+    An integer that specifies the level of debugging verbosity that we're interested in; 0 is no debugging whereas 4 is more than you probably want. If this is left undefined, it will fall back to anything set externally within the same PowerShell instance.
 
 .PARAMETER ConfigFile  
     The path to a JSON configuration file. Any parameters passed within this file will override CLI parameters EXCEPT 'Debug'. If 'Debug' is set on the CLI, it cannot be overriden by the configuration file.
@@ -39,7 +58,12 @@
         -Debug
 
 .NOTES
-
+    Author: Ben Hurwitz
+    Email: bchurwitz+pixi_pyprojSetup@gmail.com
+    Date: 2025-May-10
+    Version: 1.0.0
+    License: GNU GPL v3.0 (see above)
+    Compatability: PowerShell 5.1 and 7
 #>
 [CmdletBinding()]
 param (